Refactor: Replace react-hot-toast with react-toastify for notifications
Fix: Update WishlistPage to use productVariantId instead of productId

Refactor: Update modal imports and usage in CategoryCard, ProductCard, and SubcategoryCard components

Style: Remove unnecessary comments and improve layout in SubcategorySlider

Enhance: Update Header component to include user authentication and cart count

Chore: Comment out Header and Footer in Homepage for testing

Refactor: Simplify RootLayout by integrating LayoutContainer and ToastContainer

Add: Implement getAddress controller for fetching user addresses

Add: Create getOrderByID controller for retrieving order details by ID

Add: Generate invoice PDFs for orders 17 and 18

Add: Implement PaymentSuccessPage for displaying order confirmation

Add: Create LayoutContainer component to manage layout structure

Add: Establish AuthContext for user authentication and cart management

+++ backend/controllers/authController/logout.controller.js
@@ -1,14 +1,29 @@
 import prisma from "../../lib/db.js";
 
 const logout = async (req, res) => {
-  const { refreshToken } = req.body;
+  const refreshToken = req.cookies?.refreshToken;
+  console.log("Cookies received:", req.cookies);
+  console.log(refreshToken, "refresh Token");
 
   try {
+    if (!refreshToken) {
+      return res.status(400).json({ message: "No refresh token found" });
+    }
     await prisma.refreshToken.deleteMany({
       where: { token: refreshToken },
     });
 
-    res.clearCookie("refreshToken");
+    res.clearCookie("token", {
+      httpOnly: true,
+      sameSite: "Lax",
+      secure: process.env.NODE_ENV === "production",
+    });
+
+    res.clearCookie("refreshToken", {
+      httpOnly: true,
+      sameSite: "Lax",
+      secure: process.env.NODE_ENV === "production",
+    });
     res.status(200).json({ message: "Logged out successfully" });
   } catch (error) {
     console.error(error);
